Code Refactoring: - rename the class ProfilePicture as Design - move some methods from Duke to Design - extract method drawUpperFrame and drawLowerFrame and printTasks to improve code readability in the method printToDoList - create exception class DukeException
