+++ src/main/java/Task.java
@@ -1,4 +1,4 @@
-public class Task {
+public abstract class Task {
     protected String task;
     protected Boolean isDone;
     public static int totalTask = 0; //total number of tasks saved
